Let SearchModule not depend on SearchPageHelper.
